+++ app/src/main/java/com/kkp/buddytrainer/domain/repository/PersonRepository.kt
@@ -5,7 +5,7 @@ import kotlinx.coroutines.flow.Flow
 
 interface PersonRepository {
     suspend fun getBuddiesFromRoom() : Flow<List<Person>>
-    suspend fun getPersonFromRoom(id : Int) : Flow<Person>
+    suspend fun getPersonFromRoom(id : Long) : Flow<Person>
     suspend fun addBuddyToRoom(person: Person)
     suspend fun deleteBuddyFromRoom(person: Person)
 }
